Further seperated assessment api into seperate modules
