+++ hawc/apps/assessment/api/pagination.py
@@ -1,0 +1,5 @@
+from rest_framework.pagination import PageNumberPagination
+
+
+class DisabledPagination(PageNumberPagination):
+    page_size = None
